Rename application_kernels to kernels
